requirements.txt
python-telegram-bot==20.3
pillow
psutil
pystray
plyer
opencv-python
pyinstaller